+++ R/pca.R
@@ -12,9 +12,7 @@
 #' @export
 #'
 #' @examples
-#' mae <- hermes::multi_assay_experiment
-#' mae_data <- dataset("MAE", mae)
-#' data <- teal_data(mae_data)
+#' data <- teal_data(MAE = hermes::multi_assay_experiment)
 #' app <- init(
 #'   data = data,
 #'   modules = modules(
@@ -59,13 +57,10 @@ tm_g_pca <- function(label,
 #' @inheritParams module_arguments
 #' @export
 ui_g_pca <- function(id,
-                     data,
                      mae_name,
                      pre_output,
                      post_output) {
   ns <- NS(id)
-  mae <- data[[mae_name]]()
-  experiment_name_choices <- names(mae)
 
   tagList(
     teal.widgets::standard_layout(
@@ -76,7 +71,7 @@ ui_g_pca <- function(id,
         ###
         tags$label("Encodings", class = "text-primary"),
         helpText("Analysis of MAE:", tags$code(mae_name)),
-        experimentSpecInput(ns("experiment"), data, mae_name),
+        uiOutput(ns("experiment_ui")),
         assaySpecInput(ns("assay")),
         conditionalPanel(
           condition = "input.tab_selected == 'PCA'",
@@ -162,9 +157,13 @@ srv_g_pca <- function(id,
                       exclude_assays) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   assert_class(filter_panel_api, "FilterPanelAPI")
-  assert_class(data, "tdata")
+  checkmate::assert_class(data, "reactive")
+  checkmate::assert_class(shiny::isolate(data()), "teal_data")
 
   moduleServer(id, function(input, output, session) {
+    output$experiment_ui <- renderUI({
+      experimentSpecInput(session$ns("experiment"), data, mae_name)
+    })
     experiment <- experimentSpecServer(
       "experiment",
       data = data,
@@ -451,9 +450,7 @@ srv_g_pca <- function(id,
 #'   sample_tm_g_pca()
 #' }
 sample_tm_g_pca <- function() {
-  mae <- hermes::multi_assay_experiment
-  mae_data <- teal.data::dataset("MAE", mae)
-  data <- teal.data::teal_data(mae_data)
+  data <- teal.data::teal_data(MAE = hermes::multi_assay_experiment)
   app <- teal::init(
     data = data,
     modules = teal::modules(