+++ R/quality.R
@@ -66,9 +66,7 @@ heatmap_plot <- function(object, assay_name) {
 #' @export
 #'
 #' @examples
-#' mae <- hermes::multi_assay_experiment
-#' mae_data <- dataset("MAE", mae)
-#' data <- teal_data(mae_data)
+#' data <- teal_data(MAE = hermes::multi_assay_experiment)
 #' app <- init(
 #'   data = data,
 #'   modules = modules(
@@ -113,7 +111,6 @@ tm_g_quality <- function(label,
 #' @inheritParams module_arguments
 #' @export
 ui_g_quality <- function(id,
-                         data,
                          mae_name,
                          pre_output,
                          post_output) {
@@ -125,7 +122,7 @@ ui_g_quality <- function(id,
       ###
       tags$label("Encodings", class = "text-primary"),
       helpText("Analysis of MAE:", tags$code(mae_name)),
-      experimentSpecInput(ns("experiment"), data, mae_name),
+      uiOutput(ns("experiment_ui")),
       selectInput(
         ns("plot_type"),
         "Plot Type",
@@ -202,9 +199,13 @@ srv_g_quality <- function(id,
                           exclude_assays) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   assert_class(filter_panel_api, "FilterPanelAPI")
-  assert_class(data, "tdata")
+  checkmate::assert_class(data, "reactive")
+  checkmate::assert_class(shiny::isolate(data()), "teal_data")
 
   moduleServer(id, function(input, output, session) {
+    output$experiment_ui <- renderUI({
+      experimentSpecInput(session$ns("experiment"), data, mae_name)
+    })
     experiment <- experimentSpecServer(
       "experiment",
       data = data,
@@ -418,9 +419,7 @@ srv_g_quality <- function(id,
 #'   sample_tm_g_quality()
 #' }
 sample_tm_g_quality <- function() {
-  mae <- hermes::multi_assay_experiment
-  mae_data <- teal.data::dataset("MAE", mae)
-  data <- teal.data::teal_data(mae_data)
+  data <- teal.data::teal_data(MAE = hermes::multi_assay_experiment)
   app <- teal::init(
     data = data,
     modules = teal::modules(