+++ NEWS.md
@@ -1,8 +1,13 @@
-# teal.modules.hermes 0.1.5.9006
+# teal.modules.hermes 0.1.5.9010
 
 ### Miscellaneous
 * Added placeholders for `assaySpec`, `adtteSpec` and `geneSpec` inputs when no option is selected.
 * Disabled the select input for `assaySpec` and `adtteSpec` when there are no options available.
+* For module developers: The module gets `teal_data` object in the UI and `reactive(<teal_data>)` in the server
+
+### Enhancements
+
+* Updated the documentation and vignettes with the new way of specifying data for `teal::init()`. The `data` argument accepts `teal_data` object.
 
 # teal.modules.hermes 0.1.5